+++ src/main/java/domain/Ladder/Ladder.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.Ladder;
 
 import util.BooleanGenerator;
