+++ src/main/java/domain/Ladder/LineGenerator.java
@@ -1,5 +1,7 @@
-package domain;
+package domain.Ladder;
 
+import domain.Ladder.Line;
+import domain.Step;
 import java.util.ArrayList;
 import java.util.List;
 import util.BooleanGenerator;
